Support shortening shortcodes
Some shortcodes, such as a shortcode to hide spoilers, would look better
if they were hidden in the widget as well. Add a setting to allow the
admin to list the shortcodes that should be shortened into only the
shortcode itself. That is, "[somecode]....[/somecode]" will turn into
"SOMECODE" when rendered.