Use str_replace as final render method
Using preg_replace with the comment content as a replacement will cause
any \N or $N occurences to be replaced with nothing (because there is no
capturing). The patterns to search for when rendering the widget don't
need to be regular expressions either. Instead use str_replace so
nothing in the comment excerpt will be treated as special.